Fix laravel

+++ database/migrations/2018_06_24_172605_create_seos_table.php
@@ -18,7 +18,7 @@ class CreateSeosTable extends Migration
             $table->string('title');
             $table->string('description');
             $table->string('keyword');
-            $table->integer('status');
+            $table->bigInteger('status');
             $table->timestamps();
         });
     }
