+++ database/migrations/2018_09_11_143425_create_sociallinks_table.php
@@ -17,7 +17,7 @@ class CreateSociallinksTable extends Migration
             $table->bigIncrements('id');
             $table->string('icon');
             $table->string('url');
-            $table->integer('status');
+            $table->bigInteger('status');
             $table->timestamps();
         });
     }
